given Congress get years

+++ src/lib.rs
@@ -1,5 +1,5 @@
 #![allow(dead_code)]
 
-pub(crate) const FIRST_CONGRESS: u16 = 1789;
+pub(crate) const FIRST_CONGRESS: usize = 1789;
 
 mod legislation;
